ClangーCLにコンパイルを変更、Imguiを新しいのにした。
Clnag-CLに変えるのに修正を行いました。
戻り値のconst が必要ないとき、削除
数字を文字列に変換したいとき、テンプレートを使うように修正
Overrideなどを追加
階層のファイルパス取得にファイルシステムで取得するように追加

+++ 3DAction_Project/3DAction_Project/ConstantBufferInfo.cpp
@@ -33,7 +33,7 @@ String ConstantBufferInfo::Serialize(int spaceNumber) const
 
     // セーブする情報を作成
     saveData += SaveUtils::MakeTypeInfo(kCBufferName, m_Name, spaceNumber);
-    saveData += SaveUtils::MakeTypeInfo(kRegisterNumber, String::to_u8string(m_RegisterNumber), spaceNumber);
+    saveData += SaveUtils::MakeTypeInfo(kRegisterNumber, String::to_u8string((uint64_t)m_RegisterNumber), spaceNumber);
     saveData += SaveUtils::MakeTypeInfo(kSize, String::to_u8string((uint64_t)m_Size), spaceNumber);
 
     return saveData;
